Move ByteTime field to SPI_DEVICE_CONFIGURATION
- Update code accordingly.

+++ src/System.Device.Spi/sys_dev_spi_native_System_Device_Spi_SpiDevice.cpp
@@ -26,7 +26,7 @@ bool System_Device_IsLongRunningOperation(
     uint32_t readSize,
     bool fullDuplex,
     bool bufferIs16bits,
-    float byteTime,
+    int32_t byteTime,
     uint32_t &estimatedDurationMiliseconds)
 {
     if (bufferIs16bits)
